% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combineall.R
\name{fct_combineall}
\alias{fct_combineall}
\title{Factor Combine}
\usage{
fct_combineall(df, .f, level, ...)
}
\arguments{
\item{df}{A data set}

\item{.f}{A factor}

\item{level}{The name of the combination level}

\item{...}{The variables to sum}
}
\description{
Combining factor levels easily
}
